Extract compare button logic from Compare component
